backend/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
autobahn==23.6.1
billiard==3.6.4.0
branca==0.6.0
brotlipy==0.7.0
cachetools==5.3.0
celery==5.2.7
channels==3.0.5
channels-redis==3.4.0
constantly==15.1.0
cssselect==1.1.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.19
django-cors-headers==4.0.0
django-filter==23.2
django-redis==5.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
ffmpy==0.3.0
gast==0.4.0
grpcio==1.54.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
mpmath==1.2.1
multidict==6.0.4
munkres==1.1.4
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
oauthlib==3.2.2
pep8==1.7.1
psycopg2==2.9.6
psycopg2-binary==2.9.3
pyarrow==12.0.0
pycurl==7.45.1
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
pytest==7.1.2
queuelib==1.5.0
redis==4.4.4
requests==2.28.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
SimpleParse==2.2.4
smmap==5.0.0
sqlparse==0.4.4
starlette==0.26.1
uvicorn==0.21.1
validators==0.20.0
vine==5.0.0
whitenoise==6.2.0
wincertstore==0.2
yarl==1.9.2